Move conversion layer to kategory.kindedj
